Add files via upload
Create a survey form with Fields; First Name, Last Name, Date of Birth, Country 
(dropdown), Gender (checkbox)
, Profession, email, and mobile number. All the input 
fields are necessary to submit the form. Create two buttons Submit and Reset. Reset will 
reset the form while clicking on submit, first, it will check all the fields and necessary 
validations and then a popup will appear displaying all the selected values with labels in 
front of it. On closing the popup, the form should reset all the values. Use the following
image for reference